[workspace model] Refactor `VirtualFileUrl` update in entities using information stored in `VirtualFileIndex`
Previously update of VirtualFileUrl's in entities was with the usage of subscription on storage change. And
each time the store change there was the loop through the whole entities of the separate type. In this commit,
we try to use information stored in VirtualFileIndex and update only related entities. It allows us to remove
code related to the storage change.  But there is a complex case with LibraryRoot which is not directly an
entity field and to fix it, it was decided to create a separate delegate VirtualFileUrlLibraryRootProperty as
a hack.

GitOrigin-RevId: b657e0a3f110edc027225276ad6237fd9b4d6113