Start crash detection after loading properties
... which happened before the (possibly custom) idea.properties contents
were loaded, and resulted in initializing PathManager.ourSystemPath
too early (when those custom properties might have altered the location
of the system dir, and therefore the location of the crash reports).

This had the side effect of not really updating plugins when running
with custom config and system dirs, as after restart the IDE would
not pick up the downloaded plugin from the correct location.

Bug: 152328198
Test: manual
Change-Id: I22faf25606d3bef7747b67233b2389b1158e2314