Bug fix: virtualenv in project dir was not being auto-detected
If you open a directory for the first time (as a new project) in
pycharm/intellij, it should auto-detect the virtualenv in that directory
and set it as the project interpreter. This bug stopped that from
happening. It was introduced by https://github.com/JetBrains/intellij-community/commit/fedf5828fff10d281e865cc3883a0a7d703d1cab