[open project] removing existing module before create a new one with the same roots (IDEA-246772, IDEA-246621)
PlatformProjectConfigurator.configureProject create a dummy module, so here we may try to create a module which use the same content root (and also has the same name). In the old project model when you add a module which name is equal to name of some existing module, reference to the old module is silently replaced in the internal map (causing memory leak) but it doesn't lead to visible problems. In the new project model it fails. So now we removes old modules before adding new one with the same content roots.

GitOrigin-RevId: 158ecf1aa8b6024c012fccc9fe409525f112c444