cwm: import configs from other IDE if there are no own configs: CWM-299
At the first start of the thin client, if there are no own thin client configs to import from, try to import from existing configs of other IDEs (if they exist on the client's machine). That allows to reuse existing configs which the user got used to, instead of starting the thin client on clean default configs.

If there are configs of multiple JetBrains IDEs existing on the client's machine, then those configs are preferred which correspond to the same IDE which is started as the host IDE for this client. Otherwise any other config is selected.

(cherry picked from commit b13dd7e9681cdd833f903245b0d9018afdc02c37)

GitOrigin-RevId: 01e5273898c90d1daa745ab78bd9e46c9f9f3263