cwm: suppress restart of the thin client if vmoptions were imported: CWM-471
The restarter doesn't support restarting of the thin client, so if vmoptions are modified during the initial config import, the thin client silently exited. Until the restarter is fixed, we prohibit restarting of the thin client, especially since changing vmpotions for the thin client doesn't seem to be valuable and often scenario.

GitOrigin-RevId: ed1b37da73716e2113dc86073c35b16777dd61c3