[file-history] walk merge commit parents according to changes made to them
When a commit merges an old version of file before rename, rename in this merge commit may not be detected. In some cases, such merge commits are not removed as trivial and participate in computing file history. This may lead to a problem when parents of such commits are traversed in a specific order: if dfs walk first enters a parent where merge was not detected, file is assumed deleted in all reachable commits. Because of this, these commit do not show up in computed history at all.

In order to solve the problem, parents of a merge commit are sorted according to path: commits with the same path go first, then commits with renamed path, and last -- commits with deleted path.

IDEA-247426

(cherry picked from commit d6643e6b9c90812cdbea7733d8a87ecc07cc6f1c)

GitOrigin-RevId: bdc41a6c0df14a0756590c77fab2db34d9ba0e74