[vcs-log] lift invokeLater out of openLogEditorTab
Method openLogEditorTab used invokeLater to open created VcsLogUi instances asynchronously. This was done to avoid opening editors under "any" modality state. However, only calls from ProjectLogListener.logCreate were under "any" and in other cases invokeLater was not required. In fact, it created a problem: there was no guarantee that invokeLater was executed prior to dispose of the corresponding VcsLogManager. That would create a problem with disposing newly created but not yet opened VcsLogUi instances.

In order to solve the problem, in this commit invokeLater is lifted out of openLogEditorTab, which now always works synchronously. The expired condition is used to ensure that tabs are not created after VcsLogManager dispose. In order to not allow this invokeLater after logDisposed, clearing the reference from VcsProjectLog to VcsLogManager it moved prior to listeners call. An additional assertion is added to the invokeLater just in case.

There are consequences of this change:
1. Old VcsLogManager is going to dangle in memory until invokeLater is either is executed or dismissed. If Settings dialog is shown at the moment, this will only happen after the dialog is closed (unless plugins are unloaded, as flushQueue is invoke in this case).
2. Vcs Log tabs are not going to be recreated until the Settings dialog is closed. With the current implementation this will happen even for toolwindow tabs (if deemed necessary, toolwindow tabs could be opened separately from editor tabs).

GitOrigin-RevId: 34778e9fdad80192d5197efe811abe9a625118c0