Fix UI traverser issue when using jars
UI traversing uses directory names to guess module names.
However, when the classes are found in jars, it uses the jar
name which does not match the module name, and in fact the same
jar name is used in multiple plugins, merging searchable
options together when they should not.

This change fixes that by coming up with a 'module' name
for jars, that is unique to this jar as long as the jar
was found inside the idea home path. This change has no effect
if traverseUI is run from sources.

Change-Id: I414c6483f78b7d4a3f353e530073054e2c355876