IDEA-CR-65956: PY-35624 Don't suggest renaming magic literals on f-strings in Django projects
Since, in general, the exact value of such literals can't be known statically,
hence it's just not safe to base a refactoring on them.

I've introduced a new method PyStringLiteralExpression#isInterpolated that
detects whether a literal contains any embedded expressions. It might be
useful in other actions that don't properly take f-strings into account yet.

It's still quite dubious that this particular renamer is suggested for *any*
string literal in a Django project causing problems like PY-35808, but it will
be revised separately.

The corresponding tests have been revised to use API available in 202 branch.

GitOrigin-RevId: a2c6a3949f9110b4689707dd81c9a7b72e99beaa