RUBY-25656 ensure OrderComparator obeys Comparator requirements when comparing elements
The previous implementation around `Comparing.equal(file1, file2)` violated the requirements for a Comparator, as it was possible to reach a situation where `a < b; b < c` but `a >= c`. This could happen when 3 elements are nested under the same parent node, and 2 of them are contained within the same file but the third is not.

GitOrigin-RevId: 1eb0df3d3356601950e62f80cbf8ddeeca0bf516