IJ-CR-2632: settings repository: fix localization for messages with parameters: IDEA-247896
vararg params was passed from icsMessage() to IcsBundle.message() as is, and it was wrapped into an array because of that. Therefore the base bundle code received Object[] as the first and only parameter, and thus was calling its toString(), which is "[Ljava.lang.Object@id".

The fix is to spread varargs into array of parameters. Since spreading is not allowed for nullable varargs, the places which passed nulls were fixed.

(cherry picked from commit 7f7fa5f7ef5ae8fa51d32b47907700e11307ba96)

GitOrigin-RevId: fd881fa8f6e60d0bd8f2acdaff09ef78af88b2c5