[util] get rid of explicit reference to internal Unsafe class in ConcurrentHashMap
Use method handle to call its methods instead. This is needed to be able to compile intellij.platform.core.impl module using JDK 11 (IDEA-248086).

The change doesn't affect performance of ConcurrentHashMap much, and anyway we plan to create a variant of it which uses var handles (see IDEA-244473), so this implementation will be used only in places which stay on Java 8.

GitOrigin-RevId: d379a7c747aa720566249d31f16d9d9e0fe13ea5