KT-40212 Fail silently from createPatternTree
MalformedPatternException was not catched from createPatternTree so it would be displayed in the log, whereas this is only meant to be displayed in the UI.

Now createPatternTree succeds and returns an empty end of line comment with a PATTERN_ERROR user data. The exception is raised in KotlinValidator like regular MalformedPatternException-s. This should not cause a huge overhead because few patterns contain a PsiComment.

GitOrigin-RevId: 4372a4f74ab1055a5c82653845ac7ef4a736d731