[i18n] idea-ui - fix ProjectFormatPanel
This patch fixes the problems that might arise for i18nized IDEA instances when they save the information about the preferred project format. It introduces a new enum called StorageFormat that is basically duplicates the records from StorageScheme but also adds numeric tags to the records and converts them to localized strings. The combobox is configured with the tags and then uses a custom renderer that visualizes the tags as localized strings.

Signed-off-by: Nikita Eshkeev <nikita.eshkeev@jetbrains.com>

GitOrigin-RevId: 41bd19fc281c8eb17166aa845d3c9f7960ed5f13