[build scripts] explicitly include kotlin-stdlib library in build scripts dependencies (IDEA-249524)
This restores dependency on blockmap library removed in 7c6f2cc45af1ed3 and also explicitly adds dependency on kotlin-stdlib. This is needed because blockmap depends on kotlin-stdlib, but the library configuration excludes transitive dependencies to avoid conflicts with other modules, so we need to include it to dependencies of this module manually to add this dependency to jps-build-script-dependencies-bootstrap library.

(cherry picked from commit ec5cb71983a1e8ab69624236ebc10699b7e61b22)

GitOrigin-RevId: eecf0630055fe6b54cdd9d6359ebcd1b83db0553