[vcs-log] add or remove highlighters from the table on toggle
Keep all of the custom highlighters in the VcsLogUiImpl, add or remove them from the table when they get enabled or disabled. After a highlighter is removed, it stops getting update calls on refresh, therefore it heeds to be updated after being added back.

After this change, all of the highlighters installed on the table are considered enabled, which means highlighter code does not need to check it.

GitOrigin-RevId: adaf03815c8b4602733f0708153edb7fbe3e756a