[openapi] EA-237806 IAE: Internals.checkLength
ErrorPaneConfigurable used to call StreamEx.zip to build a stream of errors with their indices which internally checks if both indices and errors are of the same size and if not an exception occurs. ErrorPaneConfigurable limits the number of errors to be shown on the panel which caused the problem. Now the call to StreamEx.zip is replaced with StreamEx.of(errors).zipWith which doesn't complain of the streams are of different size and ends when one of the streams ends.

Signed-off-by: Nikita Eshkeev <nikita.eshkeev@jetbrains.com>

GitOrigin-RevId: 6aa5f34c8c37b79c67636d03a1f760f3ff6dfaeb