[workspace model] initialize project-level libraries before firing first rootsChanged event
Before this change if an instance of ModuleRootListener is registered in plugin.xml, it'll get rootsChanged event during processing WorkspaceModelChangeListener in ModuleManagerComponentBridge inside WorkspaceModelTopics.notifyModulesAreLoaded call. At this moment modules are initialized but project libraries may not be initialized yet (they'll be added after handling WorkspaceModelChangeListener in ProjectLibraryTableBridgeImpl), so LibraryTable::getLibraries will return empty array. It may lead to errors if some code (e.g. in DartSdk.getDartSdk) caches the results of computation based on project-level libraries and won't be updated.

Now project-level libraries are loaded together with modules before notifyModulesAreLoaded is called.

GitOrigin-RevId: d154edad2de337fbe41f3f0b2a1f74fb1f9a900b