IDEA-CR-66660: storage: fix migration of antialiasing settings: IDEA-209384
If there is a deprecated setting (e.g. ideAAType), in UISettings#migrateOldSettings at first we take it to use as the current value, and then we overwrite this deprecated value by the default value, so that at the next start it won't be used anymore, and only the value from the new storage will be considered.

However, migrateOldSettings() happens inside loadState(). And by default the State digest is calculated after loading it. Thus setting the deprecated ideAAType value to "greyscale" happens before this calculation, so this "greyscale" value is considered as the value actually written on disk. And when we save the State, it doesn't get updated, because nothing has been changed from IDEA's point of view.

The solution is to mark the State with useLoadedStateAsExisting=false, which makes the IDE to compare with the actual state on disk rather than with the loaded state.

(cherry picked from commit 721640e39b7f6a133ea355b3d598693725f1cdde)

GitOrigin-RevId: e3f931f22c873598e4aa74230533c7165f85ce87