[github] improve determine local branch name for multiple PRs from one fork
Assume that different PRs from one fork will have a different names for PR's branches. So if we didn't match local branch by PR branch name, then we should return null. Otherwise, we may mistakenly find already present branch for different PR but within the same fork.

But if the PR was made not from the fork, then chance that the user have a local branch with a different name increases. So we can try to "guess" it by returning first branch with tracking if any.

GitOrigin-RevId: 50cc64bc9277d28041b5d85797e1d1ef25924dfb