[git] Do not filter repositories which have the selected branch as current in ShowDiffWithBranchAction
With such a filtering, calling the action for current branch fails. The action in fact compares the working tree with the branch so it makes sense to call it for the current branch as well, and it will show the list of modified files. While not very useful (it is the same list as in Local Changes), having this action for current branch is consistent.

Partially reverts be5bf2802

GitOrigin-RevId: f29aae7583279cc9e0c0fffcbcd8cecb93a3b8b9