[workspace model] invert dependencies for workspaceModel.ide module
Before workspaceModel.ide worked as a plugin: it had dependencies on other platform modules and overrode platform's services to change platform's behavior. It complicates development because we were needed to provide a way to extend functionality in all places where we want to change behavior for workspace model and it wasn't possible to implement interfaces from Workspace Model in platfrom classes. Since now workspace model is enabled by default, it makes sense to invert the dependencies: now platform modules depend on workspaceModel.ide. Legacy bridge implementations are moved to lang.impl module (because Facet and ModuleRootManagerComponent are located in lang modules and it's simpler to have all implementation of project model elements in the same module), classes related to loading/storing JPS configuration in IDE are moved to configurationStore.impl module, classes which depend on other IDE classes are moved to ide.impl module, classes related to external system API are moved to externalSystem.impl module. Implementations of some classes are merged into corresponding platform classes to simplify code.

GitOrigin-RevId: b34471658871b4c9c311e7947156008fa61f2567