[configuration store] add IComponentStore::clearCaches method to force saving in tests
It's convenient to check that 'save' action for a project works properly by loading a project from a directory, removing its configuration files, invoking 'save' and checking the results. However 'save' will do nothing if project model isn't changed so we need to clear cached data in the component stores to ensure that save is really called.

GitOrigin-RevId: e17b3548c4be126d7755baae9fe54e1a35341da0