[eclipse] rewrite EclipseClasspathTest to use more high-level methods
The old implementation invokes EclipseClasspathReader/EclipseClasspathWriter directly, so the test ignores the fact that these classes aren't used in workspace model. Now it uses high-level methods to load and save project configuration so it'll verify correctness regardless of currently used implementation of the project model.

Also the test is converted to Kotlin and migrated from JUnit3 to JUnit4.

GitOrigin-RevId: f5ca7750905265f5ddd216a98838408885764505