IDEA-251003: add `IDEA with Android` run configuration to community
and remove `Android Studio` run configuration. It does not work, and it is not possible to repair it: there is no compatible kotlin plugin for IJ master that pretends to be Android Studio (kotlin plugin for IJ explicitly says that it is `<incompatible-with>com.intellij.modules.androidstudio</incompatible-with>`, latest kotlin plugin for AS explicitly declares `until-build` which is usually one major version before the master version).

GitOrigin-RevId: 8152962e29ea2a78feca66965a40b5f5c2108cb8