IDEA-218688 Refactor UpdateChecker for flexibility
* Separate the patch downloading logic into PatchDownloadTask, which
UpdateInfoDialog calls
* Add a "callback container" so callers to
UpdateChecker.updateAndShowResult can do as they choose with the
callback's fields

Previously, a call into UpdateChecker.updateAndShowResult would check
for updates to the platform, check for updates to plugins, and then show
the UpdateInfoDialog prompting the user to accept or decline the update.

With this change, the caller to updateAndShowResult passes in its own
UpdateResultCallbackProvider, which creates a custom
UpdateResultCallback with the results of platform/plugin updates. Then,
the UpdateResultCallback is run inside showUpdateResult. In the regular
flow, this callback shows UpdateInfoDialog, and clicking on Update would
start the PatchDownloadTask. In custom usages, the UpdateInfoDialog
could be skipped and directly start PatchDownloadTask.