[vcs-log] fix "withPrefix" function for filters not from the predefined set
Function "withPrefix" is used to calculate filter collection presentation when there are several filters set. It makes filter presentation more human readable by adding a prefix to the filter "displayText", but only can do so for a predefined set of filters. This commit fixes the function for other filters by returning "displayText" instead of an empty string for them.

GitOrigin-RevId: 01bf8caca64b380714c2467286094b16e8d6f245