[Build tw] kotlinc output parser "non-file" messages title fix
The arguments for details and title were in the wrong order causing only
the first line of the error to be displayed in the right hand pane of
the build tool window. While the whole stacktrace was shown as the node
title in the left hand side tree view.

(cherry picked from commit a996d9855e2da87e37e3296a3b6cd67e72197f8e)

Change-Id: Id282fd53d415d1b355820e51aa9907cce7c24604