[git-index] display operation icons on the right
Instead of using a label for operation icon, icon is painted manually. This allows to explicitly place the icon on the right border of the tree visible area, even when the node itself does not fit there. Since now the icon could be painted on top of text, a background it added in a form of ColorIcon.

GitOrigin-RevId: 527929f0aded8f6067ff87811594cc805372bd53