EA-238869 PY-43773 Make PyPackageManager automatically dispose with its SDK
It removes the risky layer of indirection when it's the responsibility of
PyPackageManagers to remove PyPackageManager instance on disposal of its SDK
and trigger its disposal. This logic still remains in case we create several
package managers for the same not-yet-disposed SDK in PyPackageManagers, but,
on the whole, it seems safer to dispose the manager as soon as its SDK becomes
invalid, not relying on any mediator for that.

GitOrigin-RevId: 9d3f76d2ab03cc529c7f40e017d5f860fa713e1e