[workspace model] IDEA-246022 Rework VirtualFileUrlManager and VirtualFileUrl to use VirtualFilePointer in IDE modules
Prior to these changes, we stored file paths in two places: VirtualFileUrlManagerImpl and in VirtualFilePointerManagerImpl.
It consumes more memory and slows down the code because we needed to create the instance of VirtualFileUrl and the instance
of VirtualFilePointer for the single URL.  From now we use our own implementation for storing paths in the storage module,
but for the IDE modules, we reuse VirtualFilePointers.
In this commit, we also add an opportunity to find an entity from the virtual file URL index which contains entity source
with a given URL.

GitOrigin-RevId: 072a144041c9c3ced78cf664b1b87a77eaaa4e04