Indexing: do not use VFS-stored content hash when calculating indexed file hash.
There is not much sense in using VFS-stored content hash of a file when calculating indexed file hash because the full file's content is always available in-memory and the hash can be calculated quickly.

The positive side of this change is that it decouples hash calculation algorithm used in the VFS and in the indexing infrastructure.

GitOrigin-RevId: 2576a5f62978f4c69c83993f14d3f93c3505915c