[project model] extract Library::getPresentableName method
Library::getName returns null for unnamed module-level libraries so it's convenient to have a method which can be used to show library in UI directly in Library interface. So the method is moved from LibraryUtil to LibraryImpl and reused in other implementations.

Also the implementation of the method is fixed to get rid of trailing '!' symbol in JAR file paths.

GitOrigin-RevId: df6c387f2859c7f12bf239e28425460d7185bb7d