[profiler] IDEA-CR-67357: support Maven test goals profiling (IDEA-251443)
The testing goals are those provided by surefire and failsafe plugins.

This is performed by passing required JVM arguments to the forked processes, as arguments for `-Dmaven.surefire.debug`. Note that non-forking executions are not supported by this contribution.

Also, at this moment pom-related properties and parameters are not considered, so any argument provided in the related pom will be ignored.

GitOrigin-RevId: a6c2d0237ef9629e5ce1743309f56fefd96c1cb9