vcs: ComplexPathVirtualFileSystem: add ability to customize VirtualFileSystem.refreshAndFindFileByPath
During IDE startup EditorsSplitters tries to reopen existing tabs (tabs which was not closed by user explicitly). During such process, the corresponding file system tries to resolve the actual VirtualFile by the given url. So to support "recreation" of particular VirtualFile the implementors should be able to provide the corresponding implementation to find or create VF by the given url.

GitOrigin-RevId: ca4fd7c9cb392c01c8fe937bc8c24833ba6f2990