RUBY-27091 Fix problem with exe paths in SDK custom configurators on macOS using zsh
macOS apps launched via GUI don't have the same environment as those launched from terminal (see EnvironmentUtil#getEnvironmentMap Javadoc for more info on this).
One of the problems this results in is inability to find an executable without its full path in certain cases.
There is a platform workaround for this in `GeneralCommandLine#createProcess`, but it did not work in this case due to our own parent env hack in `RubyLocalRunner#passParentEnv`, where `ParentEnvironmentType#NONE` is set.
I've extracted the platform workaround to `PathEnvironmentVariableUtil#clarifyExePathOnMac` and reused it `RubySystemVersionManagerHandler#doPatchCommandLine` directly.

GitOrigin-RevId: 0947ac9e53e6e11f10bf30e2b089d4c6978af424