Split folder group title and action description in PyRootTypeProvider interface
It allows them to have different text and capitalization, better matching those
for standard roots such as source roots and excluded directories. For instance,
previously it was "Source Roots" and "Mark directory as a source root" but,
awkwardly "Template Root" (in singlular) and "Template Root".

For consistency with the new method getFolderGroupTitle(), I also renamed getColor()
to getFolderGroupColor() and deprecated the old method.

All methods for customizing this UI of "Project Structure" have been documented.

GitOrigin-RevId: 3850c2f16f2bafc6ba1b4d0a684f0d116ce6bb60