[IntelliJ project configuration] remove unnecessary dependency from intellij.properties to intellij.java.resources
This dependency was added in 1ad38c29bb to allow running tests in their own classpath but now it works even without it. And this dependency add java modules to runtime classpath of all modules which depend on intellij.properties and it causes some problems e.g. in AppCode tests.

GitOrigin-RevId: 29c762a40a784d5fdc25fa7396c191ec616822c6