[platform] fire rootsChanged event for nested 'makeRootsChange' calls
Before if ProjectRootManagerImpl.makeRootsChange is called recursively (this may happen e.g. if a library is edited in the old project model implementation) it fired 'beforeRootsChanged' for the outer and the nested call, but 'rootsChanged' was fired only for the nested call. This led to incorrectly built RootIndex and massive test failures (e.g. in DirectoryIndexTest and RootsChangedTest). Now we remembers how many times roots change event is triggered in myPendingRootsChanged and invoke 'fireRootsChanged' for all of them.

GitOrigin-RevId: 21e966db47f7060273b3e367a2f1a8dfe4eb4e1a