[live-templates] Enable "HTML/XML/T" template on XML tags
It was explicitly enabled on text elements only but with dummy identifier it was available in much more contexts like any element (even unbalanced). After removing dummy element for surround, this caused failure of the SurroundWithTest.testSurroundWithTagFirstElement test, which looks logical, as that test doesn't surround a text. Now a separate 'XML tag' context is created that also matches any balanced tag.

GitOrigin-RevId: 347ae75073de7e5eb065762e1b7b8b7bd9a5a346