[java-dfa] Do not rely on class qualified name when checking for assignability/convertibility
Qualified name is absent for local classes, so it doesn't work correctly if local classes are inherited.
Also, as we already have PsiClass objects themselves, we can skip unnecessary resolve by qualified name.
Finally, we can go further in isConvertibleFrom and inline isInheritorOrSelf. This allows to deduplicate areElementsEquivalent, which is supposed to be symmetrical, so extra work is avoided.
Fixes IDEA-253169 Inheritors of local class cause false-positive 'ArrayStoreException' warning

GitOrigin-RevId: cbadd6e6e853ae2310427dd977187b61cf1b5d5e