PY-39725 Console doesn't remove prompt when copy multiline commands
The problem is that we get prompt width from highlighters, which we add during printing. But multilines commands are printed with buffer, so we attach all the highlighters to the first line on multiline command. That leads to incorrect prompt width for non-first lines in the command. We should calculate default prompt width and use it in case we didn't find highlighter for the line.

GitOrigin-RevId: 5dfae704db0ad4508bf4546baf6492ca4e70a01f