[github] improve local PR branch detection for PRs not from fork (IDEA-253172)
In case PRs coming not from a fork, the PR's remote and the current user's remote is the same. So to determine the current local branch we should firstly find PR's remote branch with the same name. Local branches with the tracking to that remote branch is a good candidates to match such PRs.

GitOrigin-RevId: d6f16893d80979e935f92c30eb8e8bd2189d7dde