reuse Gradle connections slightly less
If we attempt to make a connection with a Gradle version which is not
currently installed, `gradleHome` in the connector parameters is null,
no matter which version of Gradle is requested.  The connector will
then download and suitably install the version of Gradle requested,
but the null value of `gradleHome` forms parts of the key for looking
up existing connections.

If there is a subsequent request to make a connection with a different
Gradle version which is also not installed, the existing stored
connection (with a different Gradle version) will be re-used rather
than installing and using the requested version.

To fix this, consider ConnectorParameter objects as different (not
equals()) if either of them has a `gradleHome` that is null.