i18n: Annotate commandInterface API, used for Django and AppEngine commands
Most of such commands are generated by directly introspecting the corresponding
library (for Django see _jb_manage_tasks_provider.py helper) and, therefore,
naturally, cannot be translated. Thus, the respective fields are annotated as
@NlsSafe.

What's more, in a few cases where we provide "stubs" for some commands that we're
not able to discover dynamically, it makes no sense to translate their hardcoded
descriptions as well, so as not to introduce a discrepancy when only some commands
are translated, while the others are not. I've reverted extracting a couple of such
messages into a bundle.

GitOrigin-RevId: 5f28c6792634ac57b547209e80e6c6c8ebcea000