[properties] IDEA-248720 Sort properties files loses comments
The property parser didn't use to include the comments that are above the
first property into the PropertiesList node it resulted in the problem when
the properties get sorted and a property with a comment goes up, its
comment doesn't get included into the PropertiesList node and hence when
the old properties list is replaced with the new (sorted) properties list
the comment disappears. Now the property file parser includes all the
comments into the PropertiesList node so no comments disappear after sorting
the file.

GitOrigin-RevId: 9cc7bd015d3d3ad6f712ff6eb1de23fcd22a4dcb