jps: Extract jps.build.javac.rt.rpc module and make an artifact from it
This is required to maintain compatibility of the module with Java 1.6.
The last protobuf-java version compatible with Java 1.6 was 3.5, so
that in order to be able to update the project library version of
protobuf-java, we split the JavacRemoteProto.java into two versions:
the one running in the external process is generated using and against
protobuf 3.5, and the other running in the IDE process - the protobuf
from project libraries.

(cherry picked from commit 325de77b7d8e52f0dbf10ddf2e4a7716ddd05fc3)

IJ-CR-3773

GitOrigin-RevId: 7b61335758c160dfbd733455720a6c3bb97e2003