[workspace model] delete caches from the old project when a new project is created at the same location (IDEA-253686)
The platform doesn't delete caches, so when a new project is created at the same location and with the same name as some project which was deleted, the new project will reuse caches from the old project. In case of workspace model caches it leads to exception (see IDEA-253686), because the code expects that the newly created project has no modules. Also this may lead to problems for other caches.

So here a new method (getProjectDataPath) is introduced; all paths to different project-specific caches returned by it are located under the same directory, and that directory is cleared when we create a new project.

(cherry picked from commit 5b4023ed6b2e2e6fb5a8d63650f99e9a9645c3ca)

IDEA-CR-68041

GitOrigin-RevId: e223e556d49c4d4fa7a42668644355e45e15ca46