Fix platform registration of AndroidStudioAnalytics
We should initialize analytics to the no-op implementation in the
following cases:
  * The "IDEA" run configuration defined in the tools/idea project,
  which runs with platform prefix "Idea".
  * Various headless application starters and utilities invoked by
  platform build scripts (to reorder jars, render SVG icons, etc.)
  running in headless mode with platform prefix "AndroidStudio".
  * Tests that do not depend on the Android plugin and don't call
  AndroidStudioInitializer.

Tests depending on the Android plugin (intellij.android.core) will
initialize the real implementation via AndroidStudioInitializer. The
same applies for Android Studio, whether running from source or a
release build. In this case we want the exception to fire if we
forget to initialize analytics.

Change-Id: Ifdb7e68cef75262ee2fac0438d8e79f7fd74de1b